+++ src/pyspartn/spartntables.py
@@ -1,5 +1,5 @@
 """
-SPARTN Lookup and Decode Tables
+SPARTN Bitmask, Lookup and Decode Tables
 
 Created on 10 Feb 2023
 
@@ -9,7 +9,7 @@ Information Sourced from https://www.spartnformat.org/download/
 :author: semuadmin
 """
 
-# satellite PRN bitmask keys and lengths
+# satellite PRN bitmask keys
 SATBITMASKKEY = {
     "GPS": "SF011",
     "GLO": "SF012",
@@ -17,6 +17,7 @@ SATBITMASKKEY = {
     "BEI": "SF094",
     "QZS": "SF095",
 }
+# satellite PRN bitmask lengths (PRN values are bitmask position + 1)
 SATBITMASKLEN = {
     "SF011": [32, 44, 56, 64],
     "SF012": [24, 36, 48, 63],
@@ -25,7 +26,7 @@ SATBITMASKLEN = {
     "SF095": [10, 40, 48, 64],
 }
 
-# phase bias bitmask keys and lengths
+# phase bias bitmask keys
 PBBITMASKKEY = {
     "GPS": "SF025",
     "GLO": "SF026",
@@ -33,15 +34,60 @@ PBBITMASKKEY = {
     "BEI": "SF103",
     "QZS": "SF104",
 }
+# phase bias bitmask lengths, phase bias values
 PBBITMASKLEN = {
-    "SF025": [6, 11],
-    "SF026": [5, 9],
-    "SF102": [8, 15],
-    "SF103": [8, 15],
-    "SF104": [6, 11],
+    "SF025": (  # GPS
+        [6, 11],
+        {
+            0: "L1C",
+            1: "L2W",
+            2: "L2L",
+            3: "L5Q",
+            # 4-10: spare phase bias
+        },
+    ),
+    "SF026": (  # GLO
+        [5, 9],
+        {
+            0: "L1C",
+            1: "L2C",
+            # 2-8: spare phase bias
+        },
+    ),
+    "SF102": (  # GAL
+        [8, 15],
+        {
+            0: "L1C",
+            1: "L5Q",
+            2: "L7Q",
+            # 3-14: spare phase bias
+        },
+    ),
+    "SF103": (  # BEI
+        [8, 15],
+        {
+            0: "L2I",
+            1: "L5P",
+            2: "L7I",
+            3: "L6I",
+            4: "L1P",
+            5: "L7P",
+            6: "L8P",
+            # 7-14: spare phase bias
+        },
+    ),
+    "SF104": (  # QZS
+        [6, 11],
+        {
+            0: "L1C",
+            1: "L2L",
+            2: "L5Q",
+            # 3-10: spare phase bias
+        },
+    ),
 }
 
-# code bias bitmask lengths
+# code bias bitmask keys
 CBBITMASKKEY = {
     "GPS": "SF027",
     "GLO": "SF028",
@@ -49,12 +95,57 @@ CBBITMASKKEY = {
     "BEI": "SF106",
     "QZS": "SF107",
 }
+# code bias bitmask lengths, code bias values
 CBBITMASKLEN = {
-    "SF027": [6, 11],
-    "SF028": [5, 9],
-    "SF105": [8, 15],
-    "SF106": [8, 15],
-    "SF107": [6, 11],
+    "SF027": (  # GPS
+        [6, 11],
+        {
+            0: "C1C",
+            1: "C2W",
+            2: "C2L",
+            3: "C5Q",
+            # 4-10: spare code bias
+        },
+    ),
+    "SF028": (  # GLO
+        [5, 9],
+        {
+            0: "C1C",
+            1: "C2C",
+            # 2 to 8 : spare code bias
+        },
+    ),
+    "SF105": (  # GAL
+        [8, 15],
+        {
+            0: "C1C",
+            1: "C5Q",
+            2: "C7Q",
+            # 3-14: spare code bias
+        },
+    ),
+    "SF106": (  # BEI
+        [8, 15],
+        {
+            0: "C2I",
+            1: "C5P",
+            2: "C7I",
+            3: "C6I",
+            4: "C1P",
+            5: "C7P",
+            6: "C8P",
+            # 7-14: spare code bias
+        },
+    ),
+    "SF107": (  # QZS
+        [6, 11],
+        {
+            0: "C1C",
+            1: "C2L",
+            2: "C5Q",
+            # 3-10: spare code bias
+        },
+    ),
 }
 
 SF015_ENUM = SF022_ENUM = {
@@ -229,54 +320,4 @@ SF098_ENUM = {
     1: "CNAV2 (L1C)",
     2: "CNAV (L2C,L5)",
     # 3-7: TBD
-}
-
-SF102_ENUM = {  # rightmost bits left to right
-    0: "L1C phase bias",
-    1: "L5Q phase bias",
-    2: "L7Q phase bias",
-    # 3-14: spare phase bias
-}
-
-SF103_ENUM = {  # rightmost bits left to right
-    0: "L2I phase bias",
-    1: "L5P phase bias",
-    2: "L7I phase bias",
-    3: "L6I phase bias",
-    4: "L1P phase bias",
-    5: "L7P phase bias",
-    6: "L8P phase bias",
-    # 7-14: spare phase bias
-}
-
-SF104_ENUM = {  # rightmost bits left to right
-    0: "L1C phase bias",
-    1: "L2L phase bias",
-    2: "L5Q phase bias",
-    # 3-10: spare phase bias
-}
-
-SF105_ENUM = {  # rightmost bits left to right
-    0: "C1C code bias",
-    1: "C5Q code bias",
-    2: "C7Q code bias",
-    # 3-14: spare code bias
-}
-
-SF106_ENUM = {  # rightmost bits left to right
-    0: "C2I code bias",
-    1: "C5P code bias",
-    2: "C7I code bias",
-    3: "C6I code bias",
-    4: "C1P code bias",
-    5: "C7P code bias",
-    6: "C8P code bias",
-    # 7-14: spare code bias
-}
-
-SF107_ENUM = {  # rightmost bits left to right
-    0: "C1C code bias",
-    1: "C2L code bias",
-    2: "C5Q code bias",
-    # 3-10: spare code bias
 }